ravian-backend/ravian-backend/ravian-backend/requirements_latest.txt
fastapi
uvicorn[standard]
pydantic>=2.0.0
pydantic-settings

python-jose[cryptography]
PyJWT
redis
psycopg2-binary
sqlalchemy
alembic
asyncpg
celery[redis]
python-multipart
python-dotenv
passlib[bcrypt]
email-validator
httpx
phonenumbers
openai
PyPDF2
pdfminer.six
langchain
